refactor: Icon 이름 형식 변경
- 이름 + Icon인줄 알고 아이콘 추가했어서 이름 형식을 Icon + 이름으로 바꿨습니다.
