+++ package.json
@@ -21,7 +21,6 @@
     "react-dom": "^18.3.1",
     "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.56.3",
-    "react-icons": "^5.5.0",
     "react-router": "^7.5.3",
     "react-router-dom": "^7.6.0",
     "tailwind-merge": "^3.3.0",
